bank/requirements.txt
pycryptodome
peewee
python-decouple
